refactor: add anime store
