+++ src/player/stores/index.ts
@@ -1,2 +1,3 @@
 export * from './player/player'
 export * from './subtitles/subtitle'
+export * from './anime/anime'
