+++ src/player/stores/subtitles/subtitle.ts
@@ -1,10 +1,9 @@
 import { defineStore } from 'pinia'
-import { usePlayerStore } from '@/player/stores'
+import { usePlayerStore, useAnimeStore } from '@/player/stores'
 import { parseASS } from '@/player/assets/utils/subtitleParsing'
-import { SubtitleState, SubtitleLine, CrunchyrollMedia } from './types'
+import { SubtitleState, SubtitleLine } from './types'
 import { LANGUAGE_TYPES, SUBTITLE_ACTIONS } from '@/player/assets/utils/constants'
 
-import animes from '@/player/animes'
 import * as kuromoji from '@/player/assets/utils/kuromoji'
 import { LANGUAGES } from '@/player/assets/utils/constants'
 
@@ -17,14 +16,6 @@ export const useSubtitleStore = defineStore('subtitle', {
       native: []
     },
     current: {},
-    anime: {
-      series_id: '',
-      title: '',
-      series_title: '',
-      season_title: '',
-      episode_number: '',
-      subtitles: []
-    },
     showSubtitles: false
   }),
   getters: {
@@ -39,55 +30,15 @@ export const useSubtitleStore = defineStore('subtitle', {
     setLanguage(language: LANGUAGES) {
       this.language = language
     },
-    crunchyrollMedia(): CrunchyrollMedia {
-      const selfWindow = window as typeof window & { self: { v1config: { media: CrunchyrollMedia } } }
-      return selfWindow.self.v1config.media
-    },
-    getAnime() {
-      const { metadata, subtitles } = this.crunchyrollMedia()
-
-      const animeData = {
-        series_id: metadata.series_id,
-        title: metadata.title,
-        series_title: metadata.series_title,
-        season_title: metadata.season_title,
-        episode_number: metadata.episode_number,
-        subtitles
-      }
-
-      if (animeData.title !== this.anime.title || animeData.episode_number !== this.anime.episode_number) {
-        this.anime = animeData
-        this.load()
-      }
-
-      setTimeout(() => this.getAnime(), 2000)
-    },
-    isSupportedAnime(): boolean {
-      return animes.some(({ series_id, seasons }) =>
-        series_id === this.anime.series_id &&
-        seasons.some(({ season_title, episodes }) =>
-          season_title === this.anime.season_title &&
-          episodes.includes(Number(this.anime.episode_number))
-        )
-      )
-    },
-    fetchCurrentSeasonNumber(): number | undefined {
-      const anime = animes.find(item => item.series_id === this.anime.series_id)
-      const season = anime?.seasons.find(season => season.season_title === this.anime.season_title)
-
-      return season?.season_number
-    },
     async load() {
-      if (!this.isSupportedAnime()) {
+      const anime = useAnimeStore()
+      if (!anime.isSupportedAnime()) {
         setTimeout(() => this.crunchyrollSubtitlesHandler('add'), 1000)
         this.showSubtitles = false
         return
       }
 
-      const baseUrl = document.querySelector('#app')?.getAttribute('extension')
-      const native = this.anime.subtitles.find((item) => item.language === this.language)?.url
-      const seasonNumber = this.fetchCurrentSeasonNumber()
-      const japanese = `${baseUrl}subtitles/${this.anime.series_title}/${seasonNumber}/${this.anime.episode_number}.ass`
+      const { native, japanese } = anime.fetchSubtitles(this.language)
 
       if (!this.tokenizer) {
         this.tokenizer = await kuromoji.startKuromoji()
